"Blocking tasks" icon now only counts uncompleted tasks.
We now animate the height of the details pane in FrontSideTask so that checking off the last blocker causes the details pane to collapse nice and smooth.
This requires setting overflow-y: hidden on the details pane, which creates a new block formatting context, which means that margins no longer collapse nicely, so we're getting rid of them to keep things at nearly the same layout we had before.